requirements.txt
mysql-connector-python==9.2.0
pygame==2.6.1
